+++ s1tbx-op-utilities/src/test/java/org/esa/nest/dat/TestLinearTodB.java
@@ -83,7 +83,7 @@ public class TestLinearTodB {
         }
         band1.setData(ProductData.createInstance(intValues));
 
-        final double[] incidence_angle = new double[64];
+        final float[] incidence_angle = new float[64];
         Arrays.fill(incidence_angle, 30.0f);
         testProduct.addTiePointGrid(new TiePointGrid(OperatorUtils.TPG_INCIDENT_ANGLE, 16, 4, 0, 0, 1, 1, incidence_angle));
